requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-telegram-bot>=20.7
python-multipart>=0.0.6
python-dotenv>=1.0.0
discord.py>=2.3.0
aiohttp>=3.9.0
playwright>=1.40.0
psutil>=5.9.0
